Move more components to separate file
